requirements.txt
colorama>=0.4.6
openai>=1.14.2
python-dotenv>=1.0.0
requests>=2.31.0
